Update core_code_budgeon_zh_edit.R
I put the script into a function where you can define several arguments (included all arguments Zack had put aside in an earlier version as well as the name of the specified ROI). It works as is, but might need tweaking if we want to adjust the plots or limit the arguments. Also one line of code setting the order of the data that I believe might be better suited outside the function. An example (with 10 iterations for bootstrap) is commented out underneath the function as an example of how to run the function.